rename: Showcase 및 ColorPalette 컴포넌트를 widgets/clothes 폴더로 분리
