+++ src/widgets/clothes/ui/ColorPalette/ColorButtons/ColorButtons.style.ts
