+++ src/widgets/clothes/ui/ColorPalette/ColorButtons/ColorButtons.tsx
