+++ src/widgets/clothes/ui/ColorPalette/ColorPalette.style.ts
