+++ src/widgets/clothes/ui/Showcase/Showcase.style.ts
