+++ src/widgets/clothes/ui/Showcase/components/ClothesSlider/ClothesSlider.style.ts
